# openapi.yaml
openapi: 3.0.3
info:
  title: WikiLivee API
  version: 3.0.0
  description: >
    Wiki-редактор с совместным редактированием в реальном времени.
    Страницы, таблицы (локальное хранение в PostgreSQL), backlinks,
    collaboration (WebSocket), версионирование, якорные комментарии,
    авторизация через JWT.
  license:
    name: MIT
servers:
  - url: https://balooai.ru
    description: Production
  - url: http://localhost:8080
    description: Локальная разработка
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Регистрация нового пользователя
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
            example:
              username: alice
              password: secret123
      responses:
        '201':
          description: Пользователь создан, токен выдан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username уже занят
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Вход в систему
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
            example:
              username: alice
              password: secret123
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Неверный логин или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/me:
    get:
      operationId: getMe
      tags: [Auth]
      summary: Текущий пользователь (из токена)
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  username:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ════════════════════════════════════════════════════════
  # PAGES
  # ════════════════════════════════════════════════════════
  /api/pages:
    get:
      operationId: listPages
      tags: [Pages]
      summary: Список всех страниц
      responses:
        '200':
          description: Список страниц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createPage
      tags: [Pages]
      summary: Создать страницу (author = текущий пользователь)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePageRequest'
      responses:
        '201':
          description: Страница создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/pages/{id}:
    get:
      operationId: getPage
      tags: [Pages]
      summary: Получить страницу
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Страница
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: savePage
      tags: [Pages]
      summary: Сохранить страницу (autosave). Создаёт снапшот в page_versions
      parameters:
        - $ref: '#/components/parameters/pageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePageRequest'
      responses:
        '200':
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deletePage
      tags: [Pages]
      summary: Удалить страницу
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/pages/{id}/backlinks:
    get:
      operationId: getPageBacklinks
      tags: [Pages]
      summary: Страницы, которые ссылаются на данную (через wikiLink-блоки)
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Список обратных ссылок
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/pages/{id}/children:
    get:
      operationId: getPageChildren
      tags: [Pages]
      summary: Дочерние страницы
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Список дочерних страниц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════
  # VERSIONING
  # ════════════════════════════════════════════════════════
  /api/pages/{id}/versions:
    get:
      operationId: getPageVersions
      tags: [Versioning]
      summary: История версий страницы
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Список версий, отсортированных по убыванию
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageVersionSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/pages/{id}/versions/{version}:
    get:
      operationId: getPageVersion
      tags: [Versioning]
      summary: Получить конкретную версию
      parameters:
        - $ref: '#/components/parameters/pageId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: Версия страницы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageVersion'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/pages/{id}/versions/{version}/restore:
    post:
      operationId: restorePageVersion
      tags: [Versioning]
      summary: Откатить страницу к версии. Сохраняет новый снапшот в page_versions
      parameters:
        - $ref: '#/components/parameters/pageId'
        - $ref: '#/components/parameters/version'
      responses:
        '200':
          description: Страница восстановлена, новая версия возвращена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatePageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════
  # COMMENTS (якорные — привязаны к позиции в тексте)
  # ════════════════════════════════════════════════════════
  /api/pages/{id}/comments:
    get:
      operationId: getComments
      tags: [Comments]
      summary: Комментарии к странице (с якорными позициями)
      parameters:
        - $ref: '#/components/parameters/pageId'
      responses:
        '200':
          description: Список комментариев, отсортированных по дате создания
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: addComment
      tags: [Comments]
      summary: Добавить комментарий. anchorFrom/anchorTo — позиции ProseMirror
      parameters:
        - $ref: '#/components/parameters/pageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
            example:
              author: Alice
              text: Нужно уточнить этот момент
              anchorFrom: 42
              anchorTo: 58
              anchorText: уточнить этот момент
      responses:
        '201':
          description: Комментарий создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteComment
      tags: [Comments]
      summary: Удалить комментарий
      parameters:
        - $ref: '#/components/parameters/pageId'
        - name: commentId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: UUID комментария
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ════════════════════════════════════════════════════════
  # SEARCH
  # ════════════════════════════════════════════════════════
  /api/pages/search:
    post:
      operationId: searchPages
      tags: [Search]
      summary: Полнотекстовый поиск по страницам
      responses:
        '200':
          description: Результаты поиска
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ════════════════════════════════════════════════════════
  # GRAPH
  # ════════════════════════════════════════════════════════
  /api/pages/graph:
    get:
      operationId: getPagesGraph
      tags: [Graph]
      summary: Граф связей между страницами (wikiLink-рёбра)
      responses:
        '200':
          description: Узлы и рёбра графа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Graph'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ════════════════════════════════════════════════════════
  # TABLES (локальное хранение в PostgreSQL)
  # ════════════════════════════════════════════════════════
  /api/tables:
    get:
      operationId: listTables
      tags: [Tables]
      summary: Список всех таблиц
      responses:
        '200':
          description: Список таблиц
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TableSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: createTable
      tags: [Tables]
      summary: Создать таблицу с колонками
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTableRequest'
            example:
              name: Задачи
              columns:
                - name: Название
                  type: text
                - name: Статус
                  type: text
      responses:
        '201':
          description: Таблица создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tables/{id}:
    get:
      operationId: getTable
      tags: [Tables]
      summary: Получить таблицу со всеми колонками и строками
      parameters:
        - $ref: '#/components/parameters/tableId'
      responses:
        '200':
          description: Таблица
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateTable
      tags: [Tables]
      summary: Переименовать таблицу
      parameters:
        - $ref: '#/components/parameters/tableId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Таблица обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteTable
      tags: [Tables]
      summary: Удалить таблицу (CASCADE удаляет колонки, строки, ячейки)
      parameters:
        - $ref: '#/components/parameters/tableId'
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tables/{id}/columns:
    post:
      operationId: addColumn
      tags: [Tables]
      summary: Добавить колонку в таблицу
      parameters:
        - $ref: '#/components/parameters/tableId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddColumnRequest'
      responses:
        '201':
          description: Колонка добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableColumn'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tables/{id}/columns/{colId}:
    delete:
      operationId: deleteColumn
      tags: [Tables]
      summary: Удалить колонку (CASCADE удаляет ячейки)
      parameters:
        - $ref: '#/components/parameters/tableId'
        - name: colId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID колонки
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tables/{id}/rows:
    post:
      operationId: addRow
      tags: [Tables]
      summary: Добавить пустую строку в таблицу
      parameters:
        - $ref: '#/components/parameters/tableId'
      responses:
        '201':
          description: Строка добавлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/tables/{id}/rows/{rowId}:
    put:
      operationId: updateRow
      tags: [Tables]
      summary: Обновить ячейки строки (upsert). values — map columnId → value
      parameters:
        - $ref: '#/components/parameters/tableId'
        - name: rowId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: UUID строки
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRowRequest'
            example:
              values:
                "col-uuid-1": "Задача А"
                "col-uuid-2": "В работе"
      responses:
        '200':
          description: Строка обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableRow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteRow
      tags: [Tables]
      summary: Удалить строку
      parameters:
        - $ref: '#/components/parameters/tableId'
        - name: rowId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Удалено
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ════════════════════════════════════════════════════════
  # WEBSOCKET
  # ════════════════════════════════════════════════════════
  /ws/pages/{id}:
    get:
      operationId: connectPageWs
      tags: [Collaboration]
      summary: WebSocket для совместного редактирования
      security: []
      description: >
        Upgrade до WebSocket. После подключения сервер немедленно рассылает
        всем в комнате сообщение `presence` с обновлённым списком участников.


        **Входящие сообщения (клиент → сервер)**:

        - `update` — сохраняет страницу в БД и рассылает всем остальным

        - `cursor` — передаёт позицию текстового курсора (ProseMirror anchor/head)


        **Исходящие сообщения (сервер → клиент)**:

        - `update` — изменения страницы от другого участника

        - `cursor` — позиция курсора другого участника

        - `presence` — актуальный список участников (при join/leave)
      parameters:
        - $ref: '#/components/parameters/pageId'
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: ID пользователя. Если не указан — генерируется anon-XXXXX
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Отображаемое имя. По умолчанию "Anonymous"
        - name: color
          in: query
          required: false
          schema:
            type: string
            example: '#3b82f6'
          description: HEX цвет курсора и аватарки
      responses:
        '101':
          description: Switching Protocols — соединение установлено
        '400':
          description: Не WebSocket запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT токен выдаётся при /api/auth/login или /api/auth/register.
        Срок действия — 30 дней.
  parameters:
    pageId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: TEXT ID страницы (hex-строка 32 символа)
    tableId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID таблицы
    version:
      name: version
      in: path
      required: true
      schema:
        type: integer
      description: Номер версии
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Требуется авторизация (Bearer токен отсутствует или невалиден)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: page not found

    # ── Auth ─────────────────────────────────────────────
    AuthRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 2
          example: alice
        password:
          type: string
          minLength: 4
          format: password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT токен. Срок действия 30 дней
        userId:
          type: string
          format: uuid
        username:
          type: string

    # ── Pages ────────────────────────────────────────────
    PageSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        icon:
          type: string
          example: 📄
        parent_id:
          type: string
          nullable: true
          description: ID родительской страницы
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time
    Page:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        icon:
          type: string
          example: 📄
        parent_id:
          type: string
          nullable: true
        content:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        version:
          type: integer
        updatedAt:
          type: string
          format: date-time
    CreatePageRequest:
      type: object
      properties:
        title:
          type: string
          default: ''
        icon:
          type: string
          default: 📄
        parentId:
          type: string
          nullable: true
          description: ID родительской страницы
    UpdatePageRequest:
      type: object
      required: [title, content, version]
      properties:
        title:
          type: string
        content:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        version:
          type: integer
          description: Текущая версия клиента (для оптимистичного контроля)
    UpdatePageResponse:
      type: object
      properties:
        id:
          type: string
        version:
          type: integer
          description: Новая версия после сохранения
    Block:
      type: object
      required: [id, type]
      properties:
        id:
          type: string
        type:
          type: string
          enum:
            - paragraph
            - heading
            - list
            - table_embed
            - page_link
        content:
          type: string
          nullable: true
          description: Текстовое содержимое блока
        props:
          type: object
          additionalProperties: true
          description: >
            Дополнительные свойства блока:
            heading → { level: 1|2|3, tiptapContent: [...] };
            table_embed → { tableId: string, tableName: string };
            page_link → { targetId: string, targetTitle: string };
            list → { ordered: boolean, tiptapContent: [...] }.

    # ── Versioning ───────────────────────────────────────
    PageVersionSummary:
      type: object
      properties:
        version:
          type: integer
        savedAt:
          type: string
          format: date-time
    PageVersion:
      type: object
      properties:
        pageId:
          type: string
        version:
          type: integer
        title:
          type: string
        content:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        savedAt:
          type: string
          format: date-time

    # ── Comments ─────────────────────────────────────────
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pageId:
          type: string
        author:
          type: string
        text:
          type: string
        anchorFrom:
          type: integer
          description: Начальная позиция выделения в ProseMirror (0 = нет якоря)
        anchorTo:
          type: integer
          description: Конечная позиция выделения в ProseMirror
        anchorText:
          type: string
          description: Снапшот выделенного текста на момент создания комментария
        createdAt:
          type: string
          format: date-time
    CreateCommentRequest:
      type: object
      required: [author, text]
      properties:
        author:
          type: string
        text:
          type: string
        anchorFrom:
          type: integer
          default: 0
          description: Начальная позиция в ProseMirror (0 = без якоря)
        anchorTo:
          type: integer
          default: 0
        anchorText:
          type: string
          default: ''
          description: Выделенный текст (снапшот)

    # ── Search ───────────────────────────────────────────
    SearchResult:
      type: object
      properties:
        pageId:
          type: string
        title:
          type: string
        snippet:
          type: string
          description: Фрагмент текста с совпадением
        updatedAt:
          type: string
          format: date-time

    # ── Graph ────────────────────────────────────────────
    Graph:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/GraphEdge'
    GraphNode:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
    GraphEdge:
      type: object
      properties:
        source:
          type: string
          description: ID страницы-источника
        target:
          type: string
          description: ID страницы-цели (на кого ссылается)

    # ── WebSocket events ─────────────────────────────────
...
    TableSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Table:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        columns:
          type: array
          items:
            $ref: '#/components/schemas/TableColumn'
        rows:
          type: array
          items:
            $ref: '#/components/schemas/TableRow'
    CreateTableRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        columns:
          type: array
          items:
            $ref: '#/components/schemas/TableColumnSpec'
          default: []
    TableColumnSpec:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [text, number, date, select]
          default: text
    AddColumnRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [text, number, date, select]
          default: text
    UpdateRowRequest:
      type: object
      required: [values]
      properties:
        values:
          type: object
          additionalProperties:
            type: string
          description: Map columnId (UUID) → string value. Только указанные ячейки обновляются (upsert)
    TableColumn:
      type: object
      properties:
        id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [text, number, date, select]
        position:
          type: integer
          description: Порядковый номер колонки (0-based)
    TableRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        table_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        values:
          type: object
          additionalProperties:
            type: string
          description: Map columnId (UUID) → string value